make image pull auth optional (#12)
# Pull Request

## Checklist
- [x] Have you read Digital Catapult's [Code of
Conduct](https://github.com/digicatapult/.github/blob/main/CODE_OF_CONDUCT.md)?
- [x] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] New and existing unit tests pass locally with my changes.

## PR Type

- [x] Chore
- [x] Documentation Update

## Linked tickets

https://digicatapult.atlassian.net/browse/BRID-103

## High level description

Make authentication for pulling the base images required for the DAGs
optional. This is part of open-sourcing the repos and moving towards
dockerhub.

Steps;
1. Add another airflow variable to check if the base image needs
authentication
2. If the base image needs auth, then get the secret name which contains
the auth
3. Update the readme with new dockerhub base image names

## Detailed description

<!-- A detailed description of the feature and if possible describe how
has been implemented. -->


## Describe alternatives you've considered

<!-- A clear and concise description of the alternative solutions you've
considered. Be sure to explain why the existing customisability isn't
suitable for this feature. -->

## Operational impact

<!--- A description of any operational considerations associated with
the change. Is there anything in particular we should be looking at when
deploying the change to make sure it is working as intended. If
something goes wrong will any special actions be needed to revert the
change. -->

## Additional context

<!-- Add any other context or screenshots about the feature request
here. -->