Feature/dvc with aws conn (#16)
# Pull Request

## Checklist
- [x] Have you read Digital Catapult's [Code of
Conduct](https://github.com/digicatapult/.github/blob/main/CODE_OF_CONDUCT.md)?
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] New and existing unit tests pass locally with my changes.

## PR Type

- [x] Bug Fix
- [x] Feature
- [x] Documentation Update


## Linked tickets

<!-- If this PR is linked to a JIRA ticket or Github Issue, please
provide the ticket URL here. -->

## High level description

Necessary changes to make the aws authentication for s3 buckets to work
1) with kind cluster where access key id and secret access key are
provided 2) when the required env variables are set for the IAM based
auth in prod

## Detailed description

<!-- A detailed description of the feature and if possible describe how
has been implemented. -->


## Describe alternatives you've considered

<!-- A clear and concise description of the alternative solutions you've
considered. Be sure to explain why the existing customisability isn't
suitable for this feature. -->

## Operational impact

<!--- A description of any operational considerations associated with
the change. Is there anything in particular we should be looking at when
deploying the change to make sure it is working as intended. If
something goes wrong will any special actions be needed to revert the
change. -->

## Additional context

<!-- Add any other context or screenshots about the feature request
here. -->